@@ -6,24 +6,31 @@ import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.auth.UserProfileChangeRequest;
+import com.google.firebase.database.annotations.Nullable;
+import com.google.firebase.storage.FirebaseStorage;
+import com.google.firebase.storage.StorageReference;
+import com.google.firebase.storage.UploadTask;
 
 public class AdminSignUpActivity extends AppCompatActivity {
 
@@ -33,7 +40,8 @@ public class AdminSignUpActivity extends AppCompatActivity {
     Button btnsubmit;
     TextView newuser;
     ImageView Userpic;
-    ProgressBar progressBar;
+    StorageReference storageReference;
+    ProgressBar progressBar;LinearLayout linearLayout;
     Uri uriprofileImg;
     private FirebaseAuth mAuth;
 
@@ -42,7 +50,7 @@ public class AdminSignUpActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_admin_sign_up);
 
-
+        storageReference = FirebaseStorage.getInstance().getReference();
         mAuth = FirebaseAuth.getInstance();
         et_username = findViewById(R.id.et_username);
         etemail = findViewById(R.id.etemail);
@@ -50,9 +58,9 @@ public class AdminSignUpActivity extends AppCompatActivity {
         etCpassword = findViewById(R.id.etCpassword);
         btnsubmit = findViewById(R.id.btn_newsubmit);
         Userpic = findViewById(R.id.iv_pic);
-        // progressBar = findViewById(R.id.progbar);
+        progressBar = findViewById(R.id.progbar);
         newuser = findViewById(R.id.tv_newuser);
-
+        linearLayout = findViewById(R.id.liner);
         Userpic.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -67,10 +75,10 @@ public class AdminSignUpActivity extends AppCompatActivity {
         btnsubmit.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                final String name = et_username.getText().toString();
-                final String email = etemail.getText().toString();
-                final String password = etpassword.getText().toString();
-                final String Cpassword = etCpassword.getText().toString();
+                final String name = et_username.getText().toString().trim();
+                final String email = etemail.getText().toString().trim();
+                final String password = etpassword.getText().toString().trim();
+                final String Cpassword = etCpassword.getText().toString().trim();
                 if (name.isEmpty()) {
                     et_username.setError("enter valid email");
                     et_username.requestFocus();
@@ -91,33 +99,16 @@ public class AdminSignUpActivity extends AppCompatActivity {
                     etCpassword.setError("password not match");
                     etCpassword.requestFocus();
                     return;
-                    /*
-                    createnewuser(email, name, password);*/
-                } else {/*
-                    etCpassword.setError("password not match");
-                    etCpassword.requestFocus();*/
-
+                } else {
+                    linearLayout.setVisibility(View.GONE);
+                    progressBar.setVisibility(View.VISIBLE);
                     createnewuser(email, name, password);
-
                 }
-
                 btnsubmit.setVisibility(View.INVISIBLE);
-
             }
-
         });
-
     }
 
-    private void openGallery() {
-        //opens gallery and waint untilnuser pick up img
-
-
-        Intent i = new Intent(Intent.ACTION_GET_CONTENT);
-        i.setType("image/*");
-        startActivityForResult(i, REQUESCODE);
-
-    }
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
@@ -133,12 +124,11 @@ public class AdminSignUpActivity extends AppCompatActivity {
         }
     }
 
-
     private void checkAndRequestForPermission() {
         if (ContextCompat.checkSelfPermission(AdminSignUpActivity.this, Manifest.permission.READ_EXTERNAL_STORAGE)
                 != PackageManager.PERMISSION_GRANTED) {
             if (ActivityCompat.shouldShowRequestPermissionRationale(AdminSignUpActivity.this, Manifest.permission.READ_EXTERNAL_STORAGE)) {
-                Toast.makeText(AdminSignUpActivity.this, "plz accept permissioin", Toast.LENGTH_SHORT).show();
+                Toast.makeText(AdminSignUpActivity.this, "pleasw accept permissioin", Toast.LENGTH_SHORT).show();
             } else {
                 ActivityCompat.requestPermissions(AdminSignUpActivity.this, new String[]{Manifest.permission.READ_EXTERNAL_STORAGE}, PReqCode);
             }
@@ -147,32 +137,39 @@ public class AdminSignUpActivity extends AppCompatActivity {
         }
     }
 
+    private void openGallery() {
+        //opens gallery and waint untilnuser pick up img
+        Intent i = new Intent(Intent.ACTION_GET_CONTENT);
+        i.setType("image/*");
+        startActivityForResult(i, REQUESCODE);
+
+    }
+
     private void createnewuser(String email, final String name, String password) {
 
         mAuth.createUserWithEmailAndPassword(email, password)
-                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
+                .addOnCompleteListener(AdminSignUpActivity.this, new OnCompleteListener<AuthResult>() {
                     @Override
                     public void onComplete(@NonNull Task<AuthResult> task) {
                         if (task.isSuccessful()) {
                             // Sign in success, update UI with the signed-in user's information
                             Log.d("MSG", "createUserWithEmail:success");
-                            Toast.makeText(AdminSignUpActivity.this, "account is created.",
+                            /*Toast.makeText(AdminSignUpActivity.this, "account is created.",
                                     Toast.LENGTH_SHORT).show();
-//                            newuser.setVisibility(View.GONE);
 
-                            //  saveUserInfo(name, uriprofileImg, mAuth.getCurrentUser());
-                            finish();
-                            Intent salesman1 = new Intent(AdminSignUpActivity.this, Salesman.class);
-                            startActivity(salesman1);
+                         */ progressBar.setVisibility(View.VISIBLE);
+                            saveUserInfo(name, uriprofileImg, mAuth.getCurrentUser());
+                            //UpdateUI();
+                            /*Intent salesman1 = new Intent(AdminSignUpActivity.this, Salesman.class);
+                            startActivity(salesman1);*/
                         } else {
                             // If sign in fails, display a message to the user.
                             Log.w("MSG", "createUserWithEmail:failure", task.getException());
                             Toast.makeText(AdminSignUpActivity.this, "Authentication failed.", Toast.LENGTH_SHORT).show();
+                            progressBar.setVisibility(View.GONE);
+                            linearLayout.setVisibility(View.VISIBLE);
                             btnsubmit.setVisibility(View.VISIBLE);
-                            //   progressBar.setVisibility(View.INVISIBLE);
                         }
-
-
                         // ...
                     }
 
@@ -180,9 +177,9 @@ public class AdminSignUpActivity extends AppCompatActivity {
     }
 
 
-    /*private void saveUserInfo(final String Name, Uri uriprofileImg, final FirebaseUser currentUser) {
-
-        StorageReference mstorageRef = FirebaseStorage.getInstance().getReference().child("users_photos");
+    private void saveUserInfo(final String Name, Uri uriprofileImg, final FirebaseUser currentUser) {
+        progressBar.setVisibility(View.VISIBLE);
+        StorageReference mstorageRef = FirebaseStorage.getInstance().getReference().child("users_DP");
         final StorageReference imgFilePAth = mstorageRef.child(uriprofileImg.getLastPathSegment());
         imgFilePAth.putFile(uriprofileImg).addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
             //img uploaded sucefuly
@@ -201,8 +198,16 @@ public class AdminSignUpActivity extends AppCompatActivity {
                             @Override
                             public void onComplete(@NonNull Task<Void> task) {
                                 if (task.isSuccessful()) {
-                                    Toast.makeText(AdminSignUpActivity.this, "updated", Toast.LENGTH_SHORT).show();
-                                    UpdateUI();
+                                    UpdateUI();progressBar.setVisibility(View.VISIBLE);
+                                    Toast.makeText(AdminSignUpActivity.this, "account is created.",
+                                            Toast.LENGTH_SHORT).show();
+                                }else {
+                                    // If sign in fails, display a message to the user.
+                                    Log.w("MSG", "createUserWithEmail:failure", task.getException());
+                                    Toast.makeText(AdminSignUpActivity.this, "Authentication failed.", Toast.LENGTH_SHORT).show();
+                                    progressBar.setVisibility(View.GONE);
+                                    linearLayout.setVisibility(View.VISIBLE);
+                                    btnsubmit.setVisibility(View.VISIBLE);
                                 }
                             }
 
@@ -215,11 +220,10 @@ public class AdminSignUpActivity extends AppCompatActivity {
         });
     }
 
-    private void UpdateUI() {
-        finish();
+    private void UpdateUI() { finish();
         Intent salesman = new Intent(AdminSignUpActivity.this, Salesman.class);
         startActivity(salesman);
+
     }
-*/
 }
 
