@@ -2,7 +2,7 @@ package com.example.mysts;
 
 import android.annotation.SuppressLint;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.TextView;
 
 public class SalesmanDetailsActivity extends AppCompatActivity {
