@@ -3,7 +3,7 @@ package com.example.mysts;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.Toast;
@@ -34,23 +34,23 @@ public class SalesmanRegistrationActivity extends AppCompatActivity {
         String email = et_email.getText().toString();
         String password = et_password.getText().toString();
         SalesmanModel model = new SalesmanModel(this);
-        if (!name.isEmpty() && !mob.isEmpty() && !email.isEmpty() && !password.isEmpty()) {
-            long id = (long) model.addSalesman(name, email, password, mob, sale_cnt);
-            if (id != -1) {
-                Toast.makeText(this, "Salesman Added with id : " + id, Toast.LENGTH_LONG).show();
-                finish();
-                Intent i = new Intent(SalesmanRegistrationActivity.this, Salesman.class);
-                startActivity(i);
+        if (!name.isEmpty() && name.length() >= 3  && !email.isEmpty() && email.contains("@") && email.contains(".") && !password.isEmpty()) {
+            if (mob.length() == 10) {
+                long id = (long) model.addSalesman(name, email, password, mob, sale_cnt);
+                if (id != -1) {
+                    Toast.makeText(this, "Salesman Added with id : " + id, Toast.LENGTH_SHORT).show();
+                    finish();
+                    Intent i = new Intent(SalesmanRegistrationActivity.this, Salesman.class);
+                    startActivity(i);
 
+                } else {
+                    Toast.makeText(this, "Failed to add salesman.", Toast.LENGTH_LONG).show();
+                }
             } else {
-                Toast.makeText(this, "Failed to add salesman.", Toast.LENGTH_LONG).show();
+                et_mob.setError("enter 10 digit mobile no.");
             }
         } else {
-            Toast.makeText(this, "Do not Enter Empty Field", Toast.LENGTH_LONG).show();
+            Toast.makeText(this, "please enter valid info", Toast.LENGTH_LONG).show();
         }
-        et_email.setText("");
-        et_mob.setText("");
-        et_name.setText("");
-        et_password.setText("");
     }
 }
