@@ -2,7 +2,7 @@ package com.example.mysts;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 public class SplacescreenActivity extends AppCompatActivity {
 
